backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.5
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Caching & Background Tasks
redis==5.0.1
celery==5.3.4

# AI/ML
openai==1.3.7
sentence-transformers==2.2.2
pinecone-client==2.2.4
langchain==0.0.340

# Portfolio Analysis
PyGithub==2.1.1
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
spacy==3.7.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
